Finished the server and client connection
Added an endpoint address for the server networking. Also made the socket into listening mode and a loop that handles the client interaction. For the client I created another socket and connected the server to the client using the connect() function which then sends and receives a response from the server.